requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
aiofiles>=23.2.1
python-multipart>=0.0.9
openai-whisper>=20231117
torch>=2.1.0
transformers>=4.36.0
sentencepiece>=0.1.99
groq>=0.3.0
tqdm>=4.66.1
numpy>=1.24.0
pydub>=0.25.1
